app/requirements.txt
psycopg2
flask_sqlalchemy
pydantic
datetime
pyyaml
flask